fix: change favicon, structure progect...
