+++ index.html
@@ -2,9 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link type="image/x-icon" href="/f.ico" rel="shortcut icon">
+    <link type="Image/x-icon" href="/f.ico" rel="icon">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <title>Component React</title>
   </head>
   <body>
     <div id="root"></div>
