+++ src/data/bd.ts
@@ -1,0 +1,44 @@
+export const cardItems = [
+  {
+    id: 1,
+    title: 'photo',
+    subTitle: 'home',
+    like: 2,
+    countEye: 6,
+  },
+  {
+    id: 2,
+    title: 'picture',
+    subTitle: 'forest',
+    like: 2,
+    countEye: 8,
+  },
+  {
+    id: 3,
+    title: 'art',
+    subTitle: 'human',
+    like: 2,
+    countEye: 7,
+  },
+  {
+    id: 4,
+    title: 'photo',
+    subTitle: 'animals',
+    like: 5,
+    countEye: 6,
+  },
+  {
+    id: 5,
+    title: 'photo',
+    subTitle: 'field',
+    like: 4,
+    countEye: 6,
+  },
+  {
+    id: 6,
+    title: 'photo',
+    subTitle: 'yard',
+    like: 2,
+    countEye: 16,
+  },
+];
