+++ src/pages/About/About.tsx
@@ -3,9 +3,7 @@ import style from './about.module.css';
 
 class About extends React.Component {
   render(): React.ReactNode {
-    return (
-      <div className={style.about}>Helllo, my name is Sergey . I am studying in rs-school</div>
-    );
+    return <div className={style.about}>Hello, my name is Sergey. I am studying in RS-school</div>;
   }
 }
 export default About;
